% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psio.R
\name{psio_flag_answers_to_verify}
\alias{psio_flag_answers_to_verify}
\title{Mark selected answer(s) as \emph{not verified}.}
\usage{
psio_flag_answers_to_verify(survey_id, response_field_ids)
}
\arguments{
\item{survey_id}{Integer indicating survey id.}

\item{response_fields}{array of Integers indicating the response field ids as
seen in /surveys/:survey/review, or from the "answers > id" in entries.}
}
\description{
Mark selected answer(s) as \emph{not verified}.
}
