% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_params.R
\name{psio_get_api_key}
\alias{psio_get_api_key}
\title{Get the API key}
\usage{
psio_get_api_key()
}
\value{
A string, invisibly. Will give an error if no key is set.
}
\description{
This function should not normally be called directly, but is used by various
other functions of the \code{rpapersurvey} package.
}
