% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{psio_get_documents}
\alias{psio_get_documents}
\title{Get documents}
\usage{
psio_get_documents(survey_id, per_page = 200L)
}
\arguments{
\item{survey_id}{Integer indicating survey id.}

\item{per_page}{Integer indicating how to chunk responses.}
}
\description{
Get documents
}
