% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{psio_get_fields}
\alias{psio_get_fields}
\title{Extract fields from entries}
\usage{
psio_get_fields(entries, recoding = NULL)
}
\arguments{
\item{entries}{Result returned by \code{\link[=psio_get_entries]{psio_get_entries()}}.}

\item{recoding}{Optional \code{data.frame} for recoding field names. If not
\code{NULL}, must include columns \code{field_id} and \code{field_name_new}.}
}
\value{
A \code{data.table} having one row per field.
}
\description{
A field corresponds to an individual question in a survey, which
distinguishes the results returned by this function and by
\code{\link[=psio_get_questions]{psio_get_questions()}}.
}
