% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{psio_get_questions}
\alias{psio_get_questions}
\title{Get questions}
\usage{
psio_get_questions(survey_id)
}
\arguments{
\item{survey_id}{Integer for survey id.}
}
\value{
A \code{data.table} having one row per question.
}
\description{
Here a question could correspond to a group of single-choice fields, which
distinguishes the results returned by this function and by
\code{\link[=psio_get_fields]{psio_get_fields()}}.
}
