% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psio.R
\name{psio_get_responses_for_review}
\alias{psio_get_responses_for_review}
\title{Get Entry Responses for review
Similar to getting entries via the Data API but more compact, with extra
metadata used when verifying/reviewing answers.}
\usage{
psio_get_responses_for_review(survey_id, cache_dir = NULL, per_page = 200L)
}
\arguments{
\item{survey_id}{Integer indicating survey id.}

\item{cache_dir}{Path to directory in which to cache results.}

\item{per_page}{Integer indicating how to chunk responses.}
}
\description{
Get Entry Responses for review
Similar to getting entries via the Data API but more compact, with extra
metadata used when verifying/reviewing answers.
}
