% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_review.R
\name{psio_get_reviews}
\alias{psio_get_reviews}
\title{Get answers for review}
\usage{
psio_get_reviews(
  survey_id,
  cache_dir = NULL,
  per_page = 200L,
  max_pages = 5L,
  sort_by = c("answer_id", "entry_id", "upload_id", "question_id", "verifiable_id",
    "quality", "answer", "accepted", "updated_at"),
  sort_order = c("DESC", "ASC"),
  filters = NULL
)
}
\arguments{
\item{survey_id}{Integer for survey id.}

\item{cache_dir}{Optional string for path to directory in which to cache
results.}

\item{per_page}{Integer for how to chunk fetching of results.}

\item{max_pages}{Integer for maximum number of pages to fetch. \code{NULL}
indicates all pages, i.e., all results.}

\item{sort_by}{String for field by which to sort results.}

\item{sort_order}{String for order in which to sort results.}

\item{filters}{Optional list for returning results meeting certain criteria.}
}
\value{
A \code{data.table} having one row per answer. Does not include prefilled
responses or responses collected via web form.
}
\description{
This function uses the Review API, which formats results slightly differently
than the Data API used by \code{\link[=psio_get_entries]{psio_get_entries()}} and thus \code{\link[=psio_get_answers]{psio_get_answers()}}.
}
