% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_survey.R
\name{psio_get_surveys}
\alias{psio_get_surveys}
\title{Get survey metadata}
\usage{
psio_get_surveys(survey_id = NULL)
}
\arguments{
\item{survey_id}{Optional integer for survey id. \code{NULL} indicates all
surveys.}
}
\value{
A \code{data.table} having one row per survey.
}
\description{
Get survey metadata
}
