% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{psio_set_api_key}
\alias{psio_set_api_key}
\title{Set the API key}
\usage{
psio_set_api_key(path = NULL)
}
\arguments{
\item{path}{Path to a text file containing an API key.}
}
\description{
Set the API key
}
