% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_review.R
\name{psio_set_reviews}
\alias{psio_set_reviews}
\title{Mark responses as verified or not verified}
\usage{
psio_set_reviews(
  survey_id,
  answer_ids,
  status = c("verified", "not verified"),
  dry_run = TRUE
)
}
\arguments{
\item{survey_id}{Integer indicating survey id.}

\item{answer_ids}{Vector of integers indicating the answer id(s), as seen in
\code{\link[=psio_get_answers]{psio_get_answers()}} ('detailed') or \code{\link[=psio_get_reviews]{psio_get_reviews()}}.}

\item{status}{String indicating review status.}

\item{dry_run}{Logical indicating whether to do a dry run or the real deal.}
}
\description{
Mark responses as verified or not verified
}
