% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_params.R
\name{psio_set_team}
\alias{psio_set_team}
\title{Set the team id for API calls}
\usage{
psio_set_team(team_id = NULL)
}
\arguments{
\item{team_id}{Integer indicating the team id, as seen in \code{\link[=psio_get_user]{psio_get_user()}}.}
}
\description{
Set the team id for API calls
}
