% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psio.R
\name{psio_verify_answers}
\alias{psio_verify_answers}
\title{Mark selected answer(s) as \emph{verified}.}
\usage{
psio_verify_answers(survey_id, response_field_ids)
}
\arguments{
\item{survey_id}{Integer indicating survey id.}

\item{response_fields}{array of Integers indicating the response field ids as
/surveys/:survey/review, or from the "answers > field > id" in entries.}
}
\description{
Mark selected answer(s) as \emph{verified}.
}
