% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{psio_get_answers}
\alias{psio_get_answers}
\title{Get answers}
\usage{
psio_get_answers(entries)
}
\arguments{
\item{entries}{Result returned by \code{\link[=psio_get_entries]{psio_get_entries()}}.}
}
\description{
Get answers
}
