% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psio.R
\name{psio_get_entries}
\alias{psio_get_entries}
\title{Get entries}
\usage{
psio_get_entries(survey_id, cache_dir = NULL, per_page = 200L)
}
\arguments{
\item{survey_id}{Integer indicating survey id.}

\item{cache_dir}{Path to directory in which to cache results.}

\item{per_page}{Integer indicating how to chunk responses.}
}
\description{
Get entries
}
