% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{psio_get_fields}
\alias{psio_get_fields}
\title{Get fields}
\usage{
psio_get_fields(entries, recoding = NULL)
}
\arguments{
\item{entries}{Result returned by \code{\link[=psio_get_entries]{psio_get_entries()}}.}

\item{recoding}{\code{data.frame} with which to recode field names.}
}
\description{
Get fields
}
