% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_review.R
\name{psio_get_reviews}
\alias{psio_get_reviews}
\title{Get entry responses for review}
\usage{
psio_get_reviews(
  survey_id,
  cache_dir = NULL,
  per_page = 200L,
  max_pages = 5L,
  ...
)
}
\arguments{
\item{survey_id}{Integer indicating survey id.}

\item{cache_dir}{Path to directory in which to cache results.}

\item{per_page}{Integer indicating how to chunk responses.}

\item{max_pages}{Integer indicating maximum number of pages to fetch.}

\item{...}{Other parameters passed to the API.}
}
\description{
Similar to getting entries via the Data API but more compact, with extra
metadata used when verifying/reviewing answers.
}
