% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psio.R
\name{psio_get_surveys}
\alias{psio_get_surveys}
\title{Get survey metadata}
\usage{
psio_get_surveys(survey_id = NULL)
}
\arguments{
\item{survey_id}{Integer indicating survey id.}
}
\description{
Get survey metadata
}
