% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_review.R
\name{psio_set_answer}
\alias{psio_set_answer}
\title{Update an answer to a particular question.}
\usage{
psio_set_answer(survey_id, answer_id, answer, dry_run = TRUE)
}
\arguments{
\item{survey_id}{Integer for survey id.}

\item{answer_id}{Integer for the answer id as seen in
\code{\link[=psio_get_answers]{psio_get_answers()}} ('detailed') or \code{\link[=psio_get_reviews]{psio_get_reviews()}}.}

\item{answer}{Value for the new answer. Depending on the type of
question, could be a string, integer, or list.}

\item{dry_run}{Logical for whether to do a dry run or the real deal.}
}
\description{
Update an answer to a particular question.
}
