% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psio.R
\name{psio_set_answer}
\alias{psio_set_answer}
\title{Replace an answer to a particular question with your new response.}
\usage{
psio_set_answer(survey_id, response_field_id, answer)
}
\arguments{
\item{survey_id}{Integer indicating survey id.}

\item{response_field_id}{Integer indicating the response field id as seen in
/surveys/:survey/review, or from the "answers > id" in entries.}

\item{answer}{String | list | Integer to replace the answer with; use a list
for single/multiple choice questions}
}
\description{
Replace an answer to a particular question with your new response.
}
