% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_vars.R
\name{psio_set_api_key}
\alias{psio_set_api_key}
\title{Set the API key}
\usage{
psio_set_api_key(path = NULL)
}
\arguments{
\item{path}{String for path to a text file containing an API key. If \code{NULL},
a secure prompt will open to enter the key.}
}
\value{
Result of call to \code{\link[base:Sys.setenv]{base::Sys.setenv()}}.
}
\description{
Setting the API key is necessary to fetch any data from PaperSurvey.io.
}
